requirements.txt
fastapi
fpdf2
google-genai
groq
joblib
numpy
pandas
psycopg2-binary
pydantic
python-dotenv
python-multipart
scikit-learn==1.6.1
sqlalchemy
uvicorn
xgboost